requirements.txt
Django==1.11.15
Django==2.1
Django==3.2.5
django_rest_swagger==2.2.0
djangorestframework==3.14.0
